Modify CMakeLists to build packages, launch file and fmt formatting header. Add node_ros for realtime image processing and corresponding launch file. Modify python scripts to find the mismatched time stamps and extracting the bag files. Modifed rviz to open with the necessary topics and config.
Modified point cloud to be outputted in the camera frame instead of
fixed frame that DSOL was by default doing. This ensures proper function
with octomap which expects a point cloud to be moving with the car frame.

Add sim_time in realtime launch file to use rosbag's time for map->fixed TF publisher.
There was already TF chain from camera->fixed but the fixed frame didn't
work with octomap as the path was in the X-Z axes.
Added octomap launch file.